wip attempt to detect missing icon; doesn't help
This doesn't actually detect our error -- Resources.getIdentifier
returns a perfectly valid resource ID.  The thing is that what the
resource shrinker actually does isn't quite to *remove* the resource,
but to replace it with an empty version; for an image file like this,
that means an empty file.  So the resource ID is perfectly valid,
but when the system goes to try to inflate the notification UI and so
to open the icon, the file contents aren't a valid image.